backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alabaster==0.7.12
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
audioread==3.0.1
Babel==2.8.0
bcrypt==3.2.0
blinker==1.7.0
boto3==1.35.16
botocore==1.35.16
Brlapi==0.8.3
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
command-not-found==0.3
composio_core==0.7.15
composio_langchain==0.7.15
composio_openai==0.7.15
ConfigArgParse==1.7
contourpy==1.3.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dacite==1.8.1
dbus-python==1.2.18
decorator==5.2.1
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
dnspython==2.7.0
docopt==0.6.2
docutils==0.17.1
duplicity==0.8.21
elastic-transport==8.17.0
elasticsearch==8.16.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
fasteners==0.14.1
feedparser==6.0.8
filelock==3.18.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
future==0.18.2
gevent==23.9.1
geventhttpclient==2.3.1
gitsome==0.8.0
greenlet==3.0.3
gyp==0.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.7
imagesize==1.3.0
importlib_metadata==8.7.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keyring==23.5.0
kiwisolver==1.4.8
langchain==0.3.25
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langsmith==0.3.42
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
librosa==0.11.0
linkify-it-py==2.0.3
llvmlite==0.44.0
lockfile==0.12.2
locust==2.31.5
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
memray==1.14.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.4.4
netifaces==0.11.0
networkx==3.4.2
nose2==0.15.1
numba==0.61.0
numpy==2.1.1
numpydoc==1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
openai==1.70.0
orjson==3.10.18
packaging==23.2
paramiko==3.5.1
pesq==0.0.4
pexpect==4.8.0
pika==1.3.2
pillow==10.4.0
platformdirs==4.3.7
pluggy==1.5.0
ply==3.11
pooch==1.8.2
prompt-toolkit==3.0.28
propcache==0.3.1
protobuf==3.12.4
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.11.2
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.17.2
PyGObject==3.42.1
pyinstrument==4.7.3
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.9.0
pyRFC3339==1.1
Pysher==1.0.8
pytest==8.3.3
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.36.2
regex==2024.11.6
reportlab==3.6.8
requests==2.32.2
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.14.6
roman==3.3
rpds-py==0.25.0
s3transfer==0.10.1
scikit-learn==1.6.1
scipy==1.15.2
SecretStorage==3.3.1
semver==3.0.4
sentry-sdk==2.29.1
setproctitle==1.2.2
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sounddevice==0.5.1
soundfile==0.13.1
soxr==0.5.0.post1
Sphinx==4.3.2
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.2
sympy==1.13.1
systemd-python==234
tenacity==9.1.2
textual==0.50.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tomli==2.0.1
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
triton==3.2.0
typer==0.15.4
types-requests==2.32.0.20250515
typing-inspection==0.4.0
typing_extensions==4.13.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uc-micro-py==1.0.3
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.2
usb-creator==0.3.7
uvicorn==0.34.2
uvloop==0.21.0
wadllib==1.3.6
watchfiles==1.0.5
wcwidth==0.2.5
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.3
wget==3.2
xdg==5
xkit==0.0.0
xonsh==0.11.0
yarl==1.20.0
zipp==3.21.0
zope.event==5.0
zope.interface==6.1
zstandard==0.23.0
